changed drivemotorstate to use floats instead of int

+++ prj/RoverState/DriveMotorState.h
@@ -3,6 +3,6 @@
 #include "pub_general.h"
 
 struct DriveMotorState {
-    int steer[WHEEL_COUNT];
-    int drive[WHEEL_COUNT];
+    float steer[WHEEL_COUNT];
+    float drive[WHEEL_COUNT];
 };
